+++ src/main/java/roomescape/controller/RoomescapeController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import roomescape.model.Reservation;
-import roomescape.model.Reservations;
+import roomescape.domain.Reservation;
+import roomescape.domain.Reservations;
 import roomescape.vo.ReservationCreateVo;
 
 @Controller
